pipeline_scripts/requirements.txt
# Dependencies for data pipeline scripts

pandas>=2.2
numpy>=1.26
pyarrow>=15.0
typer[all]>=0.12
pydantic>=2.5
pyyaml>=6.0
lxml>=5.0
jsonschema>=4.19
pytest>=7.4

